Split MvRewriteOptimizationTest.java to multi files
