+++ fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvRewriteTest.java
@@ -14,613 +14,27 @@
 
 package com.starrocks.sql.optimizer.rule.transformation.materialization;
 
-import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Lists;
 import com.starrocks.catalog.BaseTableInfo;
 import com.starrocks.catalog.Column;
-import com.starrocks.catalog.Database;
 import com.starrocks.catalog.ForeignKeyConstraint;
 import com.starrocks.catalog.MaterializedView;
 import com.starrocks.catalog.OlapTable;
 import com.starrocks.catalog.Table;
 import com.starrocks.catalog.UniqueConstraint;
-import com.starrocks.common.Config;
 import com.starrocks.common.FeConstants;
-import com.starrocks.connector.hive.HiveMetaClient;
-import com.starrocks.connector.hive.MockedHiveMetadata;
-import com.starrocks.pseudocluster.PseudoCluster;
-import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.ShowResultSet;
 import com.starrocks.server.GlobalStateMgr;
-import com.starrocks.sql.analyzer.Analyzer;
-import com.starrocks.sql.ast.QueryRelation;
-import com.starrocks.sql.ast.QueryStatement;
-import com.starrocks.sql.ast.StatementBase;
 import com.starrocks.sql.optimizer.OptExpression;
-import com.starrocks.sql.optimizer.Optimizer;
-import com.starrocks.sql.optimizer.base.ColumnRefFactory;
-import com.starrocks.sql.optimizer.base.ColumnRefSet;
-import com.starrocks.sql.optimizer.base.PhysicalPropertySet;
 import com.starrocks.sql.optimizer.operator.physical.PhysicalScanOperator;
-import com.starrocks.sql.optimizer.transformer.LogicalPlan;
-import com.starrocks.sql.optimizer.transformer.RelationTransformer;
-import com.starrocks.sql.parser.ParsingException;
-import com.starrocks.sql.plan.ConnectorPlanTestBase;
 import com.starrocks.sql.plan.PlanTestBase;
-import com.starrocks.thrift.TExplainLevel;
-import com.starrocks.utframe.StarRocksAssert;
-import com.starrocks.utframe.UtFrameUtils;
-import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.Test;
 
 import java.sql.SQLException;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class MvRewriteOptimizationTest {
-    private static ConnectContext connectContext;
-    private static PseudoCluster cluster;
-    private static StarRocksAssert starRocksAssert;
-
-    @BeforeClass
-    public static void beforeClass() throws Exception {
-        Config.dynamic_partition_check_interval_seconds = 10000;
-        Config.bdbje_heartbeat_timeout_second = 60;
-        Config.bdbje_replica_ack_timeout_second = 60;
-        Config.bdbje_lock_timeout_second = 60;
-        // set some parameters to speedup test
-        Config.tablet_sched_checker_interval_seconds = 1;
-        Config.tablet_sched_repair_delay_factor_second = 1;
-        Config.enable_new_publish_mechanism = true;
-
-        PseudoCluster.getOrCreateWithRandomPort(true, 3);
-        GlobalStateMgr.getCurrentState().getTabletChecker().setInterval(1000);
-        cluster = PseudoCluster.getInstance();
-
-        connectContext = UtFrameUtils.createDefaultCtx();
-        connectContext.getSessionVariable().setOptimizerExecuteTimeout(30000000);
-        connectContext.getSessionVariable().setEnableOptimizerTraceLog(true);
-
-        ConnectorPlanTestBase.mockHiveCatalog(connectContext);
-        starRocksAssert = new StarRocksAssert(connectContext);
-        starRocksAssert.withDatabase("test").useDatabase("test");
-
-        Config.enable_experimental_mv = true;
-
-        starRocksAssert.withTable("create table emps (\n" +
-                        "    empid int not null,\n" +
-                        "    deptno int not null,\n" +
-                        "    name varchar(25) not null,\n" +
-                        "    salary double\n" +
-                        ")\n" +
-                        "distributed by hash(`empid`) buckets 10\n" +
-                        "properties (\n" +
-                        "\"replication_num\" = \"1\"\n" +
-                        ");")
-                .withTable("create table emps_par (\n" +
-                        "    empid int not null,\n" +
-                        "    deptno int not null,\n" +
-                        "    name varchar(25) not null,\n" +
-                        "    salary double\n" +
-                        ")\n" +
-                        "PARTITION BY RANGE(`deptno`)\n" +
-                        "(PARTITION p1 VALUES [(\"-2147483648\"), (\"2\")),\n" +
-                        "PARTITION p2 VALUES [(\"2\"), (\"3\")),\n" +
-                        "PARTITION p3 VALUES [(\"3\"), (\"4\")))\n" +
-                        "distributed by hash(`empid`) buckets 10\n" +
-                        "properties (\n" +
-                        "\"replication_num\" = \"1\"\n" +
-                        ");")
-                .withTable("create table depts (\n" +
-                        "    deptno int not null,\n" +
-                        "    name varchar(25) not null\n" +
-                        ")\n" +
-                        "distributed by hash(`deptno`) buckets 10\n" +
-                        "properties (\n" +
-                        "\"replication_num\" = \"1\"\n" +
-                        ");")
-                .withTable("create table dependents (\n" +
-                        "    empid int not null,\n" +
-                        "    name varchar(25) not null\n" +
-                        ")\n" +
-                        "distributed by hash(`empid`) buckets 10\n" +
-                        "properties (\n" +
-                        "\"replication_num\" = \"1\"\n" +
-                        ");")
-                .withTable("create table locations (\n" +
-                        "    empid int not null,\n" +
-                        "    name varchar(25) not null\n" +
-                        ")\n" +
-                        "distributed by hash(`empid`) buckets 10\n" +
-                        "properties (\n" +
-                        "\"replication_num\" = \"1\"\n" +
-                        ");");
-        starRocksAssert.withTable("CREATE TABLE `test_all_type` (\n" +
-                "  `t1a` varchar(20) NULL COMMENT \"\",\n" +
-                "  `t1b` smallint(6) NULL COMMENT \"\",\n" +
-                "  `t1c` int(11) NULL COMMENT \"\",\n" +
-                "  `t1d` bigint(20) NULL COMMENT \"\",\n" +
-                "  `t1e` float NULL COMMENT \"\",\n" +
-                "  `t1f` double NULL COMMENT \"\",\n" +
-                "  `t1g` bigint(20) NULL COMMENT \"\",\n" +
-                "  `id_datetime` datetime NULL COMMENT \"\",\n" +
-                "  `id_date` date NULL COMMENT \"\", \n" +
-                "  `id_decimal` decimal(10,2) NULL COMMENT \"\" \n" +
-                ") ENGINE=OLAP\n" +
-                "DUPLICATE KEY(`t1a`)\n" +
-                "COMMENT \"OLAP\"\n" +
-                "DISTRIBUTED BY HASH(`t1a`) BUCKETS 3\n" +
-                "PROPERTIES (\n" +
-                "\"replication_num\" = \"1\",\n" +
-                "\"in_memory\" = \"false\"\n" +
-                ");");
-        starRocksAssert.withTable("CREATE TABLE `t0` (\n" +
-                "  `v1` bigint NULL COMMENT \"\",\n" +
-                "  `v2` bigint NULL COMMENT \"\",\n" +
-                "  `v3` bigint NULL\n" +
-                ") ENGINE=OLAP\n" +
-                "DUPLICATE KEY(`v1`, `v2`, v3)\n" +
-                "DISTRIBUTED BY HASH(`v1`) BUCKETS 3\n" +
-                "PROPERTIES (\n" +
-                "\"replication_num\" = \"1\",\n" +
-                "\"in_memory\" = \"false\"\n" +
-                ");");
-
-        starRocksAssert.withTable("CREATE TABLE `table_with_partition` (\n" +
-                "  `t1a` varchar(20) NULL COMMENT \"\",\n" +
-                "  `id_date` date NULL COMMENT \"\", \n" +
-                "  `t1b` smallint(6) NULL COMMENT \"\",\n" +
-                "  `t1c` int(11) NULL COMMENT \"\",\n" +
-                "  `t1d` bigint(20) NULL COMMENT \"\"\n" +
-                ") ENGINE=OLAP\n" +
-                "DUPLICATE KEY(`t1a`,`id_date`)\n" +
-                "COMMENT \"OLAP\"\n" +
-                "PARTITION BY RANGE(`id_date`)\n" +
-                "(PARTITION p1991 VALUES [('1991-01-01'), ('1992-01-01')),\n" +
-                "PARTITION p1992 VALUES [('1992-01-01'), ('1993-01-01')),\n" +
-                "PARTITION p1993 VALUES [('1993-01-01'), ('1994-01-01')))" +
-                "DISTRIBUTED BY HASH(`t1a`) BUCKETS 3\n" +
-                "PROPERTIES (\n" +
-                "\"replication_num\" = \"1\",\n" +
-                "\"in_memory\" = \"false\"\n" +
-                ");");
-
-        starRocksAssert.withTable("CREATE TABLE `table_with_day_partition` (\n" +
-                "  `t1a` varchar(20) NULL COMMENT \"\",\n" +
-                "  `id_date` date NULL COMMENT \"\", \n" +
-                "  `t1b` smallint(6) NULL COMMENT \"\",\n" +
-                "  `t1c` int(11) NULL COMMENT \"\",\n" +
-                "  `t1d` bigint(20) NULL COMMENT \"\"\n" +
-                ") ENGINE=OLAP\n" +
-                "DUPLICATE KEY(`t1a`,`id_date`)\n" +
-                "COMMENT \"OLAP\"\n" +
-                "PARTITION BY RANGE(`id_date`)\n" +
-                "(PARTITION p19910330 VALUES [('1991-03-30'), ('1991-03-31')),\n" +
-                "PARTITION p19910331 VALUES [('1991-03-31'), ('1991-04-01')),\n" +
-                "PARTITION p19910401 VALUES [('1991-04-01'), ('1991-04-02')),\n" +
-                "PARTITION p19910402 VALUES [('1991-04-02'), ('1991-04-03')))" +
-                "DISTRIBUTED BY HASH(`t1a`) BUCKETS 3\n" +
-                "PROPERTIES (\n" +
-                "\"replication_num\" = \"1\"\n" +
-                ");");
-
-        starRocksAssert.withTable("create table test_base_part(c1 int, c2 bigint, c3 bigint, c4 bigint)" +
-                " partition by range(c3) (" +
-                " partition p1 values less than (\"100\")," +
-                " partition p2 values less than (\"200\")," +
-                " partition p3 values less than (\"1000\")," +
-                " PARTITION p4 values less than (\"2000\")," +
-                " PARTITION p5 values less than (\"3000\"))" +
-                " distributed by hash(c1)" +
-                " properties (\"replication_num\"=\"1\");");
-
-        cluster.runSql("test", "insert into emps values(1, 1, \"emp_name1\", 100);");
-        cluster.runSql("test", "insert into emps values(2, 1, \"emp_name1\", 120);");
-        cluster.runSql("test", "insert into emps values(3, 1, \"emp_name1\", 150);");
-        cluster.runSql("test", "insert into depts values(1, \"dept_name1\")");
-        cluster.runSql("test", "insert into depts values(2, \"dept_name2\")");
-        cluster.runSql("test", "insert into depts values(3, \"dept_name3\")");
-        cluster.runSql("test", "insert into dependents values(1, \"dependent_name1\")");
-        cluster.runSql("test", "insert into locations values(1, \"location1\")");
-        cluster.runSql("test", "insert into t0 values(1, 2, 3)");
-        cluster.runSql("test", "insert into test_all_type values(" +
-                "\"value1\", 1, 2, 3, 4.0, 5.0, 6, \"2022-11-11 10:00:01\", \"2022-11-11\", 10.12)");
-
-        starRocksAssert.withTable("CREATE TABLE `t1` (\n" +
-                "  `k1` int(11) NULL COMMENT \"\",\n" +
-                "  `v1` int(11) NULL COMMENT \"\",\n" +
-                "  `v2` int(11) NULL COMMENT \"\"\n" +
-                ") ENGINE=OLAP\n" +
-                "DUPLICATE KEY(`k1`)\n" +
-                "COMMENT \"OLAP\"\n" +
-                "PARTITION BY RANGE(`k1`)\n" +
-                "(PARTITION p1 VALUES [(\"-2147483648\"), (\"2\")),\n" +
-                "PARTITION p2 VALUES [(\"2\"), (\"3\")),\n" +
-                "PARTITION p3 VALUES [(\"3\"), (\"4\")),\n" +
-                "PARTITION p4 VALUES [(\"4\"), (\"5\")),\n" +
-                "PARTITION p5 VALUES [(\"5\"), (\"6\")),\n" +
-                "PARTITION p6 VALUES [(\"6\"), (\"7\")))\n" +
-                "DISTRIBUTED BY HASH(`k1`) BUCKETS 6\n" +
-                "PROPERTIES (\n" +
-                "\"replication_num\" = \"1\"\n" +
-                ");");
-        cluster.runSql("test", "insert into t1 values (1,1,1),(1,1,2),(1,1,3),(1,2,1),(1,2,2),(1,2,3)," +
-                " (1,3,1),(1,3,2),(1,3,3)\n" +
-                " ,(2,1,1),(2,1,2),(2,1,3),(2,2,1),(2,2,2),(2,2,3),(2,3,1),(2,3,2),(2,3,3)\n" +
-                " ,(3,1,1),(3,1,2),(3,1,3),(3,2,1),(3,2,2),(3,2,3),(3,3,1),(3,3,2),(3,3,3)");
-
-        starRocksAssert.withTable("CREATE TABLE `json_tbl` (\n" +
-                "  `p_dt` date NULL COMMENT \"\",\n" +
-                "  `d_user` json NULL COMMENT \"\"\n" +
-                ") ENGINE=OLAP\n" +
-                "DUPLICATE KEY(`p_dt`)\n" +
-                "DISTRIBUTED BY HASH(p_dt) BUCKETS 2\n" +
-                "PROPERTIES ( \"replication_num\" = \"1\");");
-        cluster.runSql("test", "insert into json_tbl values('2020-01-01', '{\"a\": 1, \"gender\": \"man\"}') ");
-        cluster.runSql("test", "insert into table_with_partition values" +
-                "(\"varchar1\", '1991-02-01', 1, 1, 1), " +
-                "(\"varchar2\",'1992-02-01', 2, 1, 1), " +
-                "(\"varchar3\", '1993-02-01', 3, 1, 1)");
-        cluster.runSql("test", "insert into table_with_day_partition values " +
-                "(\"varchar1\", '1991-03-30', 1, 1, 1)," +
-                "(\"varchar2\", '1991-03-31', 2, 1, 1), " +
-                "(\"varchar3\", '1991-04-01', 3, 1, 1)," +
-                "(\"varchar3\", '1991-04-02', 4, 1, 1)");
-    }
-
-    @AfterClass
-    public static void tearDown() throws Exception {
-        PseudoCluster.getInstance().shutdown(true);
-    }
-
-    @Ignore
-    public void testSingleTableRewrite() throws Exception {
-        testSingleTableEqualPredicateRewrite();
-        testSingleTableRangePredicateRewrite();
-        testMultiMvsForSingleTable();
-        testNestedMvOnSingleTable();
-        testSingleTableResidualPredicateRewrite();
-    }
-
-    public void testSingleTableEqualPredicateRewrite() throws Exception {
-        createAndRefreshMv("test", "mv_1",
-                "create materialized view mv_1 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid = 5");
-        String query = "select empid, deptno, name, salary from emps where empid = 5";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "1:Project\n" +
-                "  |  <slot 1> : 5: empid\n" +
-                "  |  <slot 2> : 6: deptno\n" +
-                "  |  <slot 3> : 7: name\n" +
-                "  |  <slot 4> : 8: salary\n" +
-                "  |  \n" +
-                "  0:OlapScanNode\n" +
-                "     TABLE: mv_1\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: 5: empid = 5\n" +
-                "     partitions=1/1\n" +
-                "     rollup: mv_1");
-        PlanTestBase.assertContains(plan, "tabletRatio=1/6");
-
-        String query2 = "select empid, deptno, name, salary from emps where empid = 6";
-        String plan2 = getFragmentPlan(query2);
-        PlanTestBase.assertNotContains(plan2, "mv_1");
-
-        String query3 = "select empid, deptno, name, salary from emps where empid > 5";
-        String plan3 = getFragmentPlan(query3);
-        PlanTestBase.assertNotContains(plan3, "mv_1");
-
-        String query4 = "select empid, deptno, name, salary from emps where empid < 5";
-        String plan4 = getFragmentPlan(query4);
-        PlanTestBase.assertNotContains(plan4, "mv_1");
-
-        String query5 = "select empid, length(name), (salary + 1) * 2 from emps where empid = 5";
-        String plan5 = getFragmentPlan(query5);
-        PlanTestBase.assertContains(plan5, "1:Project\n" +
-                "  |  <slot 1> : 7: empid\n" +
-                "  |  <slot 5> : length(9: name)\n" +
-                "  |  <slot 6> : 10: salary + 1.0 * 2.0\n" +
-                "  |  \n" +
-                "  0:OlapScanNode\n" +
-                "     TABLE: mv_1\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: 7: empid = 5\n" +
-                "     partitions=1/1\n" +
-                "     rollup: mv_1\n" +
-                "     tabletRatio=1/6");
-
-        String query7 = "select empid, deptno from emps where empid = 5";
-        String plan7 = getFragmentPlan(query7);
-        PlanTestBase.assertContains(plan7, "mv_1");
-        OptExpression optimizedPlan7 = getOptimizedPlan(query7, connectContext);
-        List<PhysicalScanOperator> scanOperators = getScanOperators(optimizedPlan7, "mv_1");
-        Assert.assertEquals(1, scanOperators.size());
-        // column prune
-        Assert.assertFalse(scanOperators.get(0).getColRefToColumnMetaMap().keySet().toString().contains("name"));
-        Assert.assertFalse(scanOperators.get(0).getColRefToColumnMetaMap().keySet().toString().contains("salary"));
-
-        connectContext.getSessionVariable().setEnableMaterializedViewRewrite(false);
-        String query6 = "select empid, deptno, name, salary from emps where empid = 5";
-        String plan6 = getFragmentPlan(query6);
-        PlanTestBase.assertNotContains(plan6, "mv_1");
-
-        dropMv("test", "mv_1");
-        connectContext.getSessionVariable().setEnableMaterializedViewRewrite(true);
-    }
-
-    @Test
-    public void testHiveSingleTableEqualPredicateRewrite() throws Exception {
-        createAndRefreshMv("test", "hive_mv_1",
-                "create materialized view hive_mv_1 distributed by hash(s_suppkey) " +
-                        "PROPERTIES (\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\"\n" +
-                        ") " +
-                        " as select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey = 5");
-        String query = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey = 5";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_mv_1");
-
-        String query2 = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey = 6";
-        String plan2 = getFragmentPlan(query2);
-        PlanTestBase.assertNotContains(plan2, "hive_mv_1");
-
-        String query3 = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey > 5";
-        String plan3 = getFragmentPlan(query3);
-        PlanTestBase.assertNotContains(plan3, "hive_mv_1");
-
-        String query4 = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey < 5";
-        String plan4 = getFragmentPlan(query4);
-        PlanTestBase.assertNotContains(plan4, "hive_mv_1");
-
-        String query5 =
-                "select s_suppkey, s_name, s_address, (s_acctbal + 1) * 2 from hive0.tpch.supplier where s_suppkey = 5";
-        String plan5 = getFragmentPlan(query5);
-        PlanTestBase.assertContains(plan5, "hive_mv_1");
-
-        dropMv("test", "hive_mv_1");
-    }
-
-    public void testSingleTableRangePredicateRewrite() throws Exception {
-        starRocksAssert.getCtx().getSessionVariable().setEnableMaterializedViewUnionRewrite(false);
-        createAndRefreshMv("test", "mv_1",
-                "create materialized view mv_1 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid < 5");
-        String query = "select empid, deptno, name, salary from emps where empid < 5";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "1:Project\n" +
-                "  |  <slot 1> : 5: empid\n" +
-                "  |  <slot 2> : 6: deptno\n" +
-                "  |  <slot 3> : 7: name\n" +
-                "  |  <slot 4> : 8: salary\n" +
-                "  |  \n" +
-                "  0:OlapScanNode\n" +
-                "     TABLE: mv_1\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: 5: empid < 5\n" +
-                "     partitions=1/1\n" +
-                "     rollup: mv_1\n" +
-                "     tabletRatio=6/6");
-
-        String query2 = "select empid, deptno, name, salary from emps where empid < 4";
-        String plan2 = getFragmentPlan(query2);
-        PlanTestBase.assertContains(plan2, "1:Project\n" +
-                "  |  <slot 1> : 5: empid\n" +
-                "  |  <slot 2> : 6: deptno\n" +
-                "  |  <slot 3> : 7: name\n" +
-                "  |  <slot 4> : 8: salary\n" +
-                "  |  \n" +
-                "  0:OlapScanNode\n" +
-                "     TABLE: mv_1\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: 5: empid <= 3");
-
-        String query3 = "select empid, deptno, name, salary from emps where empid <= 5";
-        String plan3 = getFragmentPlan(query3);
-        PlanTestBase.assertNotContains(plan3, "mv_1");
-
-        String query4 = "select empid, deptno, name, salary from emps where empid > 5";
-        String plan4 = getFragmentPlan(query4);
-        PlanTestBase.assertNotContains(plan4, "mv_1");
-
-        String query5 = "select empid, length(name), (salary + 1) * 2 from emps where empid = 4";
-        String plan5 = getFragmentPlan(query5);
-        PlanTestBase.assertContains(plan5, "1:Project\n" +
-                "  |  <slot 1> : 7: empid\n" +
-                "  |  <slot 5> : length(9: name)\n" +
-                "  |  <slot 6> : 10: salary + 1.0 * 2.0\n" +
-                "  |  \n" +
-                "  0:OlapScanNode\n" +
-                "     TABLE: mv_1\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: 7: empid = 4");
-
-        String query6 = "select empid, length(name), (salary + 1) * 2 from emps where empid between 3 and 4";
-        String plan6 = getFragmentPlan(query6);
-        PlanTestBase.assertContains(plan6, "1:Project\n" +
-                "  |  <slot 1> : 7: empid\n" +
-                "  |  <slot 5> : length(9: name)\n" +
-                "  |  <slot 6> : 10: salary + 1.0 * 2.0\n" +
-                "  |  \n" +
-                "  0:OlapScanNode\n" +
-                "     TABLE: mv_1\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: 7: empid <= 4, 7: empid >= 3");
-
-        String query7 = "select empid, length(name), (salary + 1) * 2 from emps where empid < 5 and salary > 100";
-        String plan7 = getFragmentPlan(query7);
-        PlanTestBase.assertContains(plan7, "1:Project\n" +
-                "  |  <slot 1> : 7: empid\n" +
-                "  |  <slot 5> : length(9: name)\n" +
-                "  |  <slot 6> : 10: salary + 1.0 * 2.0\n" +
-                "  |  \n" +
-                "  0:OlapScanNode\n" +
-                "     TABLE: mv_1\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: 10: salary > 100.0");
-        dropMv("test", "mv_1");
-
-        createAndRefreshMv("test", "mv_2",
-                "create materialized view mv_2 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid < 5 and salary > 100");
-        String query8 = "select empid, length(name), (salary + 1) * 2 from emps where empid < 5";
-        String plan8 = getFragmentPlan(query8);
-        PlanTestBase.assertNotContains(plan8, "mv_2");
-
-        String query9 = "select empid, length(name), (salary + 1) * 2 from emps where empid < 5 and salary > 90";
-        String plan9 = getFragmentPlan(query9);
-        PlanTestBase.assertNotContains(plan9, "mv_2");
-        dropMv("test", "mv_2");
-
-        createAndRefreshMv("test", "mv_3",
-                "create materialized view mv_3 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid < 5 or salary > 100");
-        String query10 = "select empid, length(name), (salary + 1) * 2 from emps where empid < 5";
-        String plan10 = getFragmentPlan(query10);
-        PlanTestBase.assertContains(plan10, "mv_3");
-
-        String query11 = "select empid, length(name), (salary + 1) * 2 from emps where empid < 5 or salary > 100";
-        String plan11 = getFragmentPlan(query11);
-        PlanTestBase.assertContains(plan11, "mv_3");
-
-        String query12 = "select empid, length(name), (salary + 1) * 2 from emps" +
-                " where empid < 5 or salary > 100 or salary < 10";
-        String plan12 = getFragmentPlan(query12);
-        PlanTestBase.assertNotContains(plan12, "mv_3");
-        dropMv("test", "mv_3");
-    }
-
-    @Test
-    public void testHiveSingleTableRangePredicateRewrite() throws Exception {
-        starRocksAssert.getCtx().getSessionVariable().setEnableMaterializedViewUnionRewrite(false);
-        createAndRefreshMv("test", "hive_mv_1",
-                "create materialized view hive_mv_1 distributed by hash(s_suppkey) " +
-                        "PROPERTIES (\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\"\n" +
-                        ") " +
-                        " as select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey < 5");
-        String query = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey < 5";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_mv_1");
-
-        String query2 = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey < 4";
-        String plan2 = getFragmentPlan(query2);
-        PlanTestBase.assertContains(plan2, "hive_mv_1");
-
-        String query3 = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey <= 5";
-        String plan3 = getFragmentPlan(query3);
-        PlanTestBase.assertNotContains(plan3, "hive_mv_1");
-
-        String query4 = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey > 5";
-        String plan4 = getFragmentPlan(query4);
-        PlanTestBase.assertNotContains(plan4, "hive_mv_1");
-
-        String query5 = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey = 4";
-        String plan5 = getFragmentPlan(query5);
-        PlanTestBase.assertContains(plan5, "hive_mv_1");
-
-        String query6 =
-                "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey between 3 and 4";
-        String plan6 = getFragmentPlan(query6);
-        PlanTestBase.assertContains(plan6, "hive_mv_1");
-
-        String query7 =
-                "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where s_suppkey < 5 and " +
-                        "s_acctbal > 100.0";
-        String plan7 = getFragmentPlan(query7);
-        PlanTestBase.assertContains(plan7, "hive_mv_1");
-
-        dropMv("test", "hive_mv_1");
-    }
-
-    public void testSingleTableResidualPredicateRewrite() throws Exception {
-        createAndRefreshMv("test", "mv_1",
-                "create materialized view mv_1 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where name like \"%abc%\" and salary * deptno > 100");
-        String query =
-                "select empid, deptno, name, salary from emps where salary * deptno > 100 and name like \"%abc%\"";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "mv_1");
-        dropMv("test", "mv_1");
-    }
-
-    @Test
-    public void testHiveSingleTableResidualPredicateRewrite() throws Exception {
-        createAndRefreshMv("test", "hive_mv_1",
-                "create materialized view hive_mv_1 distributed by hash(s_suppkey) " +
-                        "PROPERTIES (\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\"\n" +
-                        ") " +
-                        " as select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where " +
-                        "s_suppkey * s_acctbal > 100 and s_name like \"%abc%\"");
-        String query = "select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier where " +
-                "s_suppkey * s_acctbal > 100 and s_name like \"%abc%\"";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_mv_1");
-        dropMv("test", "hive_mv_1");
-    }
-
-    public void testMultiMvsForSingleTable() throws Exception {
-        createAndRefreshMv("test", "mv_1",
-                "create materialized view mv_1 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid < 5");
-        createAndRefreshMv("test", "mv_2",
-                "create materialized view mv_2 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid < 6 and salary > 100");
-        String query = "select empid, length(name), (salary + 1) * 2 from emps where empid < 3 and salary > 110";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "mv_");
-        dropMv("test", "mv_1");
-        dropMv("test", "mv_2");
-
-        createAndRefreshMv("test", "mv_1",
-                "create materialized view mv_1 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid < 5");
-        createAndRefreshMv("test", "mv_2",
-                "create materialized view mv_2 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid < 5 and salary > 100");
-        String query1 = "select empid, length(name), (salary + 1) * 2 from emps where empid < 5 and salary > 110";
-        String plan2 = getFragmentPlan(query1);
-        PlanTestBase.assertContains(plan2, "mv_");
-        dropMv("test", "mv_1");
-        dropMv("test", "mv_2");
-
-        createAndRefreshMv("test", "agg_mv_1",
-                "create materialized view agg_mv_1 distributed by hash(empid)" +
-                        " as select empid, deptno, sum(salary) as total_salary from emps" +
-                        " where empid < 5 group by empid, deptno");
-        createAndRefreshMv("test", "agg_mv_2",
-                "create materialized view agg_mv_2 distributed by hash(empid)" +
-                        " as select empid, deptno, sum(salary) as total_salary from emps" +
-                        " where empid < 10 group by empid, deptno");
-        String query2 = "select empid, sum(salary) from emps where empid < 5 group by empid";
-        String plan3 = getFragmentPlan(query2);
-        PlanTestBase.assertContains(plan3, "agg_mv_");
-        dropMv("test", "agg_mv_1");
-        dropMv("test", "agg_mv_2");
-    }
-
-    public void testNestedMvOnSingleTable() throws Exception {
-        createAndRefreshMv("test", "mv_1",
-                "create materialized view mv_1 distributed by hash(empid)" +
-                        " as select empid, deptno, name, salary from emps where empid < 5");
-        createAndRefreshMv("test", "mv_2",
-                "create materialized view mv_2 distributed by hash(empid)" +
-                        " as select empid, deptno, salary from mv_1 where salary > 100");
-        String query = "select empid, deptno, (salary + 1) * 2 from emps where empid < 5 and salary > 110";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "TABLE: mv_2\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: 8: salary > 110.0, 6: empid <= 4\n" +
-                "     partitions=1/1\n" +
-                "     rollup: mv_2");
-        dropMv("test", "mv_1");
-        dropMv("test", "mv_2");
-    }
-
+public class MvRewriteTest extends MvRewriteTestBase {
     @Test
     public void testViewBasedMv() throws Exception {
         {
@@ -998,7 +412,6 @@ public class MvRewriteOptimizationTest {
         String query1 = "SELECT (test_all_type.t1d + 1) * 2, test_all_type.t1c" +
                 " from t0 join test_all_type on t0.v1 = test_all_type.t1d where t0.v1 < 100";
         String plan1 = getFragmentPlan(query1);
-        System.out.println(plan1);
         PlanTestBase.assertContains(plan1, "cross_join_mv1");
         dropMv("test", "cross_join_mv1");
 
@@ -1864,507 +1277,6 @@ public class MvRewriteOptimizationTest {
     }
 
     @Test
-    public void testNestedMv() throws Exception {
-        starRocksAssert.withTable("CREATE TABLE nest_base_table_1 (\n" +
-                "    k1 INT,\n" +
-                "    v1 INT,\n" +
-                "    v2 INT)\n" +
-                "DUPLICATE KEY(k1)\n" +
-                "PARTITION BY RANGE(`k1`)\n" +
-                "(\n" +
-                "PARTITION `p1` VALUES LESS THAN ('2'),\n" +
-                "PARTITION `p2` VALUES LESS THAN ('3'),\n" +
-                "PARTITION `p3` VALUES LESS THAN ('4'),\n" +
-                "PARTITION `p4` VALUES LESS THAN ('5'),\n" +
-                "PARTITION `p5` VALUES LESS THAN ('6'),\n" +
-                "PARTITION `p6` VALUES LESS THAN ('7')\n" +
-                ")\n" +
-                "DISTRIBUTED BY HASH(k1);");
-        cluster.runSql("test", "insert into t1 values (1,1,1),(1,1,2),(1,1,3),(1,2,1),(1,2,2),(1,2,3)," +
-                " (1,3,1),(1,3,2),(1,3,3)\n" +
-                " ,(2,1,1),(2,1,2),(2,1,3),(2,2,1),(2,2,2),(2,2,3),(2,3,1),(2,3,2),(2,3,3)\n" +
-                " ,(3,1,1),(3,1,2),(3,1,3),(3,2,1),(3,2,2),(3,2,3),(3,3,1),(3,3,2),(3,3,3);");
-        createAndRefreshMv("test", "nested_mv_1", "CREATE MATERIALIZED VIEW nested_mv_1" +
-                " PARTITION BY k1 DISTRIBUTED BY HASH(k1) BUCKETS 10\n" +
-                "REFRESH MANUAL AS SELECT k1, v1 as k2, v2 as k3 from t1;");
-        createAndRefreshMv("test", "nested_mv_1", "CREATE MATERIALIZED VIEW nested_mv_2 " +
-                "PARTITION BY k1 DISTRIBUTED BY HASH(k1) BUCKETS 10\n" +
-                "REFRESH MANUAL AS SELECT k1, count(k2) as count_k2, sum(k3) as sum_k3 from nested_mv_1 group by k1;");
-        starRocksAssert.withMaterializedView("CREATE MATERIALIZED VIEW nested_mv_3 DISTRIBUTED BY HASH(k1)\n" +
-                "REFRESH MANUAL AS SELECT k1, count_k2, sum_k3 from nested_mv_2 where k1 >1;");
-        cluster.runSql("test", "insert into t1 values (4,1,1);");
-        refreshMaterializedView("test", "nested_mv_1");
-        refreshMaterializedView("test", "nested_mv_2");
-        String query1 = "SELECT k1, count(v1), sum(v2) from t1 where k1 >1 group by k1";
-        String plan1 = getFragmentPlan(query1);
-        PlanTestBase.assertNotContains(plan1, "nested_mv_3");
-
-        dropMv("test", "nested_mv_1");
-        dropMv("test", "nested_mv_2");
-        dropMv("test", "nested_mv_3");
-        starRocksAssert.dropTable("nest_base_table_1");
-    }
-
-    @Test
-    public void testPartialPartition1() throws Exception {
-        createAndRefreshMv("test", "partial_mv",
-                "create materialized view partial_mv" +
-                        " partition by id_date" +
-                        " distributed by hash(`t1a`)" +
-                        " as" +
-                        " select t1a, id_date, t1b from table_with_partition");
-        // modify p1991 and make it outdated
-        // so p1992 and p1993 are updated
-        cluster.runSql("test", "insert into table_with_partition partition(p1991)" +
-                " values(\"varchar12\", '1991-03-01', 2, 1, 1)");
-
-        String query = "select t1a, id_date, t1b from table_with_partition" +
-                " where id_date >= '1993-02-01' and id_date < '1993-05-01'";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "partial_mv");
-
-        String query2 = "select t1a, id_date, t1b from table_with_partition" +
-                " where id_date >= '1992-01-01' and id_date < '1993-01-01'";
-        String plan2 = getFragmentPlan(query2);
-        PlanTestBase.assertContains(plan2, "partial_mv");
-        PlanTestBase.assertNotContains(plan2, "PREDICATES:");
-
-        dropMv("test", "partial_mv");
-    }
-
-    @Test
-    public void testPartialPartition2() throws Exception {
-        createAndRefreshMv("test", "partial_mv_2",
-                "create materialized view partial_mv_2" +
-                        " partition by id_date" +
-                        " distributed by hash(`t1a`)" +
-                        " as" +
-                        " select t1a, id_date, t1b from table_with_partition where t1b > 100");
-        cluster.runSql("test", "insert into table_with_partition partition(p1991)" +
-                " values(\"varchar12\", '1991-03-01', 2, 1, 1)");
-        String query4 = "select t1a, id_date, t1b from table_with_partition" +
-                " where t1b > 110 and id_date >= '1993-02-01' and id_date < '1993-05-01'";
-        String plan4 = getFragmentPlan(query4);
-        PlanTestBase.assertContains(plan4, "partial_mv_2");
-        dropMv("test", "partial_mv_2");
-    }
-
-
-    @Test
-    public void testPartialPartition3_1() throws Exception {
-        createAndRefreshMv("test", "partial_mv_3_1",
-                "create materialized view partial_mv_3_1" +
-                        " partition by date_trunc('month', new_date)" +
-                        " distributed by hash(`t1a`)" +
-                        " as" +
-                        " select t1a, id_date as new_date, t1b from table_with_day_partition");
-        cluster.runSql("test", "insert into table_with_day_partition partition(p19910331)" +
-                " values(\"varchar12\", '1991-03-31', 2, 2, 1)");
-        String query5 = "select t1a, id_date, t1b from table_with_day_partition" +
-                " where id_date >= '1991-04-01' and id_date < '1991-04-03'";
-        String plan5 = getFragmentPlan(query5);
-        PlanTestBase.assertContains(plan5, "partial_mv_3_1");
-        dropMv("test", "partial_mv_3_1");
-    }
-
-
-    @Test
-    public void testPartialPartition3_2() throws Exception {
-        createAndRefreshMv("test", "partial_mv_3_2",
-                "create materialized view partial_mv_3_2" +
-                        " partition by new_date" +
-                        " distributed by hash(`t1a`)" +
-                        " as" +
-                        " select t1a, date_trunc('month', id_date) as new_date, t1b from table_with_day_partition");
-        cluster.runSql("test", "insert into table_with_day_partition partition(p19910331)" +
-                " values(\"varchar12\", '1991-03-31', 2, 2, 1)");
-        String query6 = "select t1a, date_trunc('month', id_date), t1b from table_with_day_partition" +
-                " where id_date >= '1991-04-01' and id_date < '1991-04-03'";
-        String plan6 = getFragmentPlan(query6);
-        PlanTestBase.assertContains(plan6, "partial_mv_3_2");
-        dropMv("test", "partial_mv_3_2");
-    }
-
-    @Test
-    public void testPartialPartition4() throws Exception {
-        createAndRefreshMv("test", "partial_mv_4",
-                "create materialized view partial_mv_4" +
-                        " partition by new_name" +
-                        " distributed by hash(`t1a`)" +
-                        " as" +
-                        " select t1a, id_date as new_name, t1b from table_with_partition");
-        cluster.runSql("test", "insert into table_with_partition partition(p1991)" +
-                " values(\"varchar12\", '1991-03-01', 2, 1, 1)");
-        String query7 = "select t1a, id_date, t1b from table_with_partition" +
-                " where id_date >= '1993-02-01' and id_date < '1993-05-01'";
-        String plan7 = getFragmentPlan(query7);
-        PlanTestBase.assertContains(plan7, "partial_mv_4");
-        dropMv("test", "partial_mv_4");
-    }
-
-    @Test
-    public void testPartialPartition5() throws Exception {
-        cluster.runSql("test", "insert into test_base_part values (1, 1, 1, 1);");
-        createAndRefreshMv("test", "partial_mv_5", "create materialized view partial_mv_5" +
-                " partition by c3" +
-                " distributed by hash(c1) as" +
-                " select c1, c3, sum(c2) as c2 from test_base_part group by c1, c3;");
-        cluster.runSql("test", "alter table test_base_part add partition p6 values less than (\"4000\")");
-        cluster.runSql("test", "insert into test_base_part partition(p6) values (1, 2, 4500, 4)");
-        String query8 = "select c3, sum(c2) from test_base_part group by c3";
-        String plan8 = getFragmentPlan(query8);
-        PlanTestBase.assertContains(plan8, "partial_mv_5");
-        PlanTestBase.assertContains(plan8, "UNION");
-        PlanTestBase.assertNotContains(plan8, "c3 < -9223372036854775808");
-
-        String query9 = "select sum(c3) from test_base_part";
-        String plan9 = getFragmentPlan(query9);
-        PlanTestBase.assertNotContains(plan9, "partial_mv_5");
-        dropMv("test", "partial_mv_5");
-    }
-
-    @Test
-    public void testPartialPartition6() throws Exception {
-        // test partition prune
-        createAndRefreshMv("test", "partial_mv_6", "create materialized view partial_mv_6" +
-                " partition by c3" +
-                " distributed by hash(c1) as" +
-                " select c1, c3, c2 from test_base_part where c3 < 2000;");
-        String query10 = "select c1, c3, c2 from test_base_part";
-        String plan10 = getFragmentPlan(query10);
-        PlanTestBase.assertContains(plan10, "partial_mv_6", "UNION", "TABLE: test_base_part\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     PREDICATES: (10: c3 >= 2000) OR (10: c3 IS NULL)\n" +
-                "     partitions=3/6\n" +
-                "     rollup: test_base_part");
-
-        String query12 = "select c1, c3, c2 from test_base_part where c3 < 2000";
-        String plan12 = getFragmentPlan(query12);
-        PlanTestBase.assertContains(plan12, "partial_mv_6");
-
-        String query13 = "select c1, c3, c2 from test_base_part where c3 < 1000";
-        String plan13 = getFragmentPlan(query13);
-        PlanTestBase.assertContains(plan13, "partial_mv_6", "c3 <= 999");
-        dropMv("test", "partial_mv_6");
-    }
-
-    @Test
-    public void testPartialPartition7() throws Exception {
-        // test bucket prune
-        createAndRefreshMv("test", "partial_mv_7", "create materialized view partial_mv_7" +
-                " partition by c3" +
-                " distributed by hash(c1) as" +
-                " select c1, c3, c2 from test_base_part where c3 < 2000 and c1 = 1;");
-        String query11 = "select c1, c3, c2 from test_base_part";
-        String plan11 = getFragmentPlan(query11);
-        PlanTestBase.assertContains(plan11, "partial_mv_7", "UNION", "TABLE: test_base_part");
-        dropMv("test", "partial_mv_7");
-    }
-
-    @Test
-    public void testPartialPartition8() throws Exception {
-        createAndRefreshMv("test", "partial_mv_8", "create materialized view partial_mv_8" +
-                " partition by c3" +
-                " distributed by hash(c1) as" +
-                " select c1, c3, c2 from test_base_part where c3 < 1000;");
-        String query14 = "select c1, c3, c2 from test_base_part where c3 < 1000";
-        String plan14 = getFragmentPlan(query14);
-        PlanTestBase.assertContains(plan14, "partial_mv_8");
-        dropMv("test", "partial_mv_8");
-    }
-
-    @Test
-    public void testPartialPartition9() throws Exception {
-        createAndRefreshMv("test", "partial_mv_9", "CREATE MATERIALIZED VIEW partial_mv_9" +
-                " PARTITION BY k1 DISTRIBUTED BY HASH(k1) BUCKETS 10\n" +
-                "REFRESH MANUAL AS SELECT k1, v1 as k2, v2 as k3 from t1;");
-        // create nested mv based on partial_mv_9
-        createAndRefreshMv("test", "partial_mv_10", "CREATE MATERIALIZED VIEW partial_mv_10" +
-                " PARTITION BY k1 DISTRIBUTED BY HASH(k1) BUCKETS 10\n" +
-                "REFRESH MANUAL AS SELECT k1, count(k2) as count_k2, sum(k3) as sum_k3 from partial_mv_9 group by k1;");
-        cluster.runSql("test", "insert into t1 values (4,1,1);");
-
-        // first refresh nest mv partial_mv_10, will do nothing
-        refreshMaterializedView("test", "partial_mv_10");
-        // then refresh mv partial_mv_9
-        refreshMaterializedView("test", "partial_mv_9");
-        String query15 = "SELECT k1, count(v1), sum(v2) from t1 group by k1";
-        String plan15 = getFragmentPlan(query15);
-        // it should be union
-        PlanTestBase.assertContains(plan15, "partial_mv_9");
-        PlanTestBase.assertNotContains(plan15, "partial_mv_10");
-        dropMv("test", "partial_mv_9");
-        dropMv("test", "partial_mv_10");
-    }
-
-    @Test
-    public void testPartialPartition10() throws Exception {
-        starRocksAssert.withTable("CREATE TABLE ttl_base_table (\n" +
-                "                            k1 INT,\n" +
-                "                            v1 INT,\n" +
-                "                            v2 INT)\n" +
-                "                        DUPLICATE KEY(k1)\n" +
-                "                        PARTITION BY RANGE(`k1`)\n" +
-                "                        (\n" +
-                "                        PARTITION `p1` VALUES LESS THAN ('2'),\n" +
-                "                        PARTITION `p2` VALUES LESS THAN ('3'),\n" +
-                "                        PARTITION `p3` VALUES LESS THAN ('4'),\n" +
-                "                        PARTITION `p4` VALUES LESS THAN ('5'),\n" +
-                "                        PARTITION `p5` VALUES LESS THAN ('6'),\n" +
-                "                        PARTITION `p6` VALUES LESS THAN ('7')\n" +
-                "                        )\n" +
-                "                        DISTRIBUTED BY HASH(k1) properties('replication_num'='1');");
-        cluster.runSql("test", "insert into ttl_base_table values (1,1,1),(1,1,2),(1,2,1),(1,2,2),\n" +
-                "                                              (2,1,1),(2,1,2),(2,2,1),(2,2,2),\n" +
-                "                                              (3,1,1),(3,1,2),(3,2,1),(3,2,2);");
-        createAndRefreshMv("test", "ttl_mv_2", "CREATE MATERIALIZED VIEW ttl_mv_2\n" +
-                "               PARTITION BY k1\n" +
-                "               DISTRIBUTED BY HASH(k1) BUCKETS 10\n" +
-                "               REFRESH ASYNC\n" +
-                "               PROPERTIES(\n" +
-                "               \"partition_ttl_number\"=\"4\"\n" +
-                "               )\n" +
-                "               AS SELECT k1, sum(v1) as sum_v1 FROM ttl_base_table group by k1;");
-        MaterializedView ttlMv2 = getMv("test", "ttl_mv_2");
-        GlobalStateMgr.getCurrentState().getDynamicPartitionScheduler().runOnceForTest();
-        Assert.assertEquals(4, ttlMv2.getPartitions().size());
-
-        String query16 = "select k1, sum(v1) FROM ttl_base_table where k1=3 group by k1";
-        String plan16 = getFragmentPlan(query16);
-        PlanTestBase.assertContains(plan16, "ttl_mv_2");
-        dropMv("test", "ttl_mv_2");
-        starRocksAssert.dropTable("ttl_base_table");
-    }
-
-    @Test
-    public void testPartialPartition11() throws Exception {
-        starRocksAssert.withTable("CREATE TABLE ttl_base_table_2 (\n" +
-                "                            k1 date,\n" +
-                "                            v1 INT,\n" +
-                "                            v2 INT)\n" +
-                "                        DUPLICATE KEY(k1)\n" +
-                "                        PARTITION BY RANGE(`k1`)\n" +
-                "                        (\n" +
-                "                        PARTITION `p1` VALUES LESS THAN ('2020-01-01'),\n" +
-                "                        PARTITION `p2` VALUES LESS THAN ('2020-02-01'),\n" +
-                "                        PARTITION `p3` VALUES LESS THAN ('2020-03-01'),\n" +
-                "                        PARTITION `p4` VALUES LESS THAN ('2020-04-01'),\n" +
-                "                        PARTITION `p5` VALUES LESS THAN ('2020-05-01'),\n" +
-                "                        PARTITION `p6` VALUES LESS THAN ('2020-06-01')\n" +
-                "                        )\n" +
-                "                        DISTRIBUTED BY HASH(k1) properties('replication_num'='1');");
-        cluster.runSql("test", "insert into ttl_base_table_2 values " +
-                " (\"2019-01-01\",1,1),(\"2019-01-01\",1,2),(\"2019-01-01\",2,1),(\"2019-01-01\",2,2),\n" +
-                " (\"2020-01-11\",1,1),(\"2020-01-11\",1,2),(\"2020-01-11\",2,1),(\"2020-01-11\",2,2),\n" +
-                " (\"2020-02-11\",1,1),(\"2020-02-11\",1,2),(\"2020-02-11\",2,1),(\"2020-02-11\",2,2);");
-        createAndRefreshMv("test", "ttl_mv_3", "CREATE MATERIALIZED VIEW ttl_mv_3\n" +
-                "               PARTITION BY k1\n" +
-                "               DISTRIBUTED BY HASH(k1) BUCKETS 10\n" +
-                "               REFRESH MANUAL\n" +
-                "               AS SELECT k1, sum(v1) as sum_v1 FROM ttl_base_table_2 group by k1;");
-        String query17 = "select k1, sum(v1) FROM ttl_base_table_2 where k1 = '2020-02-11' group by k1";
-        String plan17 = getFragmentPlan(query17);
-        PlanTestBase.assertContains(plan17, "ttl_mv_3", "k1 = '2020-02-11'");
-        dropMv("test", "ttl_mv_3");
-        starRocksAssert.dropTable("ttl_base_table_2");
-    }
-
-    @Test
-    public void testHivePartialPartitionWithTTL() throws Exception {
-        starRocksAssert.getCtx().getSessionVariable().setEnableMaterializedViewUnionRewrite(true);
-        createAndRefreshMv("test", "hive_parttbl_mv",
-                "CREATE MATERIALIZED VIEW `hive_parttbl_mv`\n" +
-                        "COMMENT \"MATERIALIZED_VIEW\"\n" +
-                        "PARTITION BY (`l_shipdate`)\n" +
-                        "DISTRIBUTED BY HASH(`l_orderkey`) BUCKETS 10\n" +
-                        "REFRESH MANUAL\n" +
-                        "PROPERTIES (\n" +
-                        "\"replication_num\" = \"1\",\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\",\n" +
-                        "\"partition_ttl_number\" = \"3\"\n" +
-                        ")\n" +
-                        "AS SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` as a;");
-
-        MaterializedView ttlMv = getMv("test", "hive_parttbl_mv");
-        GlobalStateMgr.getCurrentState().getDynamicPartitionScheduler().runOnceForTest();
-        Assert.assertEquals(3, ttlMv.getPartitions().size());
-
-        String query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par`";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "0:UNION");
-
-        query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` " +
-                "where l_shipdate = '1998-01-01'";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "HdfsScanNode");
-
-        query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` " +
-                "where l_shipdate >= '1998-01-04'";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv");
-
-        dropMv("test", "hive_parttbl_mv");
-    }
-
-    @Test
-    public void testHivePartialPartition() throws Exception {
-        starRocksAssert.getCtx().getSessionVariable().setEnableMaterializedViewUnionRewrite(true);
-        createAndRefreshMv("test", "hive_parttbl_mv",
-                "CREATE MATERIALIZED VIEW `hive_parttbl_mv`\n" +
-                        "COMMENT \"MATERIALIZED_VIEW\"\n" +
-                        "PARTITION BY (`l_shipdate`)\n" +
-                        "DISTRIBUTED BY HASH(`l_orderkey`) BUCKETS 10\n" +
-                        "REFRESH MANUAL\n" +
-                        "PROPERTIES (\n" +
-                        "\"replication_num\" = \"1\",\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\",\n" +
-                        "\"storage_medium\" = \"HDD\"\n" +
-                        ")\n" +
-                        "AS SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` as a;");
-
-        String query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par`";
-        String plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv");
-
-        MockedHiveMetadata mockedHiveMetadata = (MockedHiveMetadata) connectContext.getGlobalStateMgr().getMetadataMgr().
-                getOptionalMetadata(MockedHiveMetadata.MOCKED_HIVE_CATALOG_NAME).get();
-        mockedHiveMetadata.updatePartitions("partitioned_db", "lineitem_par",
-                ImmutableList.of("l_shipdate=" + HiveMetaClient.PARTITION_NULL_VALUE));
-
-        query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` " +
-                "where l_shipdate > '1998-01-04'";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv");
-
-        query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` " +
-                "where l_shipdate > '1998-01-04' and l_shipdate < '1998-01-06'";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv");
-
-        query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` ";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv", "UNION", "PARTITION PREDICATES: ((22: l_shipdate < '1998-01-01')" +
-                " OR (22: l_shipdate >= '1998-01-06')) OR (22: l_shipdate IS NULL)");
-        dropMv("test", "hive_parttbl_mv");
-
-        createAndRefreshMv("test", "hive_parttbl_mv_2",
-                "CREATE MATERIALIZED VIEW `hive_parttbl_mv_2`\n" +
-                        "COMMENT \"MATERIALIZED_VIEW\"\n" +
-                        "PARTITION BY (`l_shipdate`)\n" +
-                        "DISTRIBUTED BY HASH(`l_orderkey`) BUCKETS 10\n" +
-                        "REFRESH MANUAL\n" +
-                        "PROPERTIES (\n" +
-                        "\"replication_num\" = \"1\",\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\",\n" +
-                        "\"storage_medium\" = \"HDD\"\n" +
-                        ")\n" +
-                        "AS SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` " +
-                        "where l_orderkey > 100;");
-        query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` " +
-                "where l_orderkey > 100;";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv_2");
-
-        mockedHiveMetadata.updatePartitions("partitioned_db", "lineitem_par",
-                ImmutableList.of("l_shipdate=1998-01-02"));
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv_2", "l_orderkey > 100", "lineitem_par",
-                "PARTITION PREDICATES: ((23: l_shipdate >= '1998-01-02') AND ((23: l_shipdate < '1998-01-03')" +
-                        " OR (23: l_shipdate >= '1998-01-06'))) OR (23: l_shipdate IS NULL)",
-                "NON-PARTITION PREDICATES: 21: l_orderkey > 100");
-
-        dropMv("test", "hive_parttbl_mv_2");
-
-        // test partition prune
-        createAndRefreshMv("test", "hive_parttbl_mv_3",
-                "CREATE MATERIALIZED VIEW `hive_parttbl_mv_3`\n" +
-                        "COMMENT \"MATERIALIZED_VIEW\"\n" +
-                        "PARTITION BY (`l_shipdate`)\n" +
-                        "DISTRIBUTED BY HASH(`l_orderkey`) BUCKETS 10\n" +
-                        "REFRESH MANUAL\n" +
-                        "PROPERTIES (\n" +
-                        "\"replication_num\" = \"1\",\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\",\n" +
-                        "\"storage_medium\" = \"HDD\"\n" +
-                        ")\n" +
-                        "AS SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` " +
-                        "where l_shipdate > '1998-01-02';");
-        query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` ";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv_3", "partitions=3/6", "lineitem_par");
-        PlanTestBase.assertNotContains(plan, "partitions=2/6");
-        dropMv("test", "hive_parttbl_mv_3");
-
-        createAndRefreshMv("test", "hive_parttbl_mv_4",
-                "CREATE MATERIALIZED VIEW `hive_parttbl_mv_4`\n" +
-                        "COMMENT \"MATERIALIZED_VIEW\"\n" +
-                        "PARTITION BY (`l_shipdate`)\n" +
-                        "DISTRIBUTED BY HASH(`l_orderkey`) BUCKETS 10\n" +
-                        "REFRESH MANUAL\n" +
-                        "PROPERTIES (\n" +
-                        "\"replication_num\" = \"1\",\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\",\n" +
-                        "\"storage_medium\" = \"HDD\"\n" +
-                        ")\n" +
-                        "AS SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` " +
-                        "where l_shipdate < '1998-01-02' and l_orderkey = 100;");
-        query = "SELECT `l_orderkey`, `l_suppkey`, `l_shipdate`  FROM `hive0`.`partitioned_db`.`lineitem_par` ";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv_4", "partitions=1/6", "lineitem_par",
-                "NON-PARTITION PREDICATES: (((22: l_shipdate < '1998-01-01') OR (22: l_shipdate >= '1998-01-02')) " +
-                        "OR (20: l_orderkey != 100)) OR (22: l_shipdate IS NULL)");
-        dropMv("test", "hive_parttbl_mv_4");
-
-        createAndRefreshMv("test", "hive_parttbl_mv_5",
-                "CREATE MATERIALIZED VIEW `hive_parttbl_mv_5`\n" +
-                        "COMMENT \"MATERIALIZED_VIEW\"\n" +
-                        "PARTITION BY date_trunc('month', o_orderdate)\n" +
-                        "DISTRIBUTED BY HASH(`o_orderkey`) BUCKETS 10\n" +
-                        "REFRESH MANUAL\n" +
-                        "PROPERTIES (\n" +
-                        "\"replication_num\" = \"1\",\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\",\n" +
-                        "\"storage_medium\" = \"HDD\"\n" +
-                        ")\n" +
-                        "AS SELECT `o_orderkey`, `o_orderstatus`, `o_orderdate`  FROM `hive0`.`partitioned_db`.`orders`");
-        query = "SELECT `o_orderkey`, `o_orderstatus`, `o_orderdate`  FROM `hive0`.`partitioned_db`.`orders`";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv_5", "360/360");
-
-        query = "SELECT `o_orderkey`, `o_orderstatus`, `o_orderdate`  FROM `hive0`.`partitioned_db`.`orders` " +
-                "where o_orderdate >= '1991-01-01' and o_orderdate < '1991-02-1'";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv_5", "partitions=1/36");
-
-        mockedHiveMetadata.updatePartitions("partitioned_db", "orders",
-                ImmutableList.of("o_orderdate=1991-02-02"));
-
-        query = "SELECT `o_orderkey`, `o_orderstatus`, `o_orderdate`  FROM `hive0`.`partitioned_db`.`orders` ";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv_5", "orders",
-                "PARTITION PREDICATES: (((15: o_orderdate < '1991-01-01') OR (15: o_orderdate >= '1991-02-01')) AND" +
-                        " ((15: o_orderdate < '1991-03-01') OR (15: o_orderdate >= '1993-12-31')))" +
-                        " OR (15: o_orderdate IS NULL)");
-
-        // TODO(Ken Huang): This should support query rewrite
-        query = "SELECT `o_orderkey`, `o_orderstatus`, `o_orderdate`  FROM `hive0`.`partitioned_db`.`orders` " +
-                "where o_orderdate >= '1992-05-01' and o_orderdate < '1992-05-31'";
-        plan = getFragmentPlan(query);
-        System.out.println(plan);
-
-        refreshMaterializedView("test", "hive_parttbl_mv_5");
-        mockedHiveMetadata.updatePartitions("partitioned_db", "orders",
-                ImmutableList.of("o_orderdate=1991-01-02"));
-        query = "SELECT `o_orderkey`, `o_orderstatus`, `o_orderdate`  FROM `hive0`.`partitioned_db`.`orders` " +
-                "where o_orderdate >= '1992-05-01' and o_orderdate < '1992-05-31'";
-        plan = getFragmentPlan(query);
-        PlanTestBase.assertContains(plan, "hive_parttbl_mv_5", "PREDICATES: 12: o_orderdate < '1992-05-31'",
-                "partitions=1/36");
-
-        dropMv("test", "hive_parttbl_mv_5");
-    }
-
-    @Test
     public void testCardinality() throws Exception {
         try {
             FeConstants.USE_MOCK_DICT_MANAGER = true;
@@ -2379,6 +1291,7 @@ public class MvRewriteOptimizationTest {
             FeConstants.USE_MOCK_DICT_MANAGER = false;
         }
     }
+
     @Test
     public void testPkFk() throws SQLException {
         cluster.runSql("test", "CREATE TABLE test.parent_table1(\n" +
@@ -2491,135 +1404,6 @@ public class MvRewriteOptimizationTest {
         Assert.assertNull(foreignKeyConstraints3);
     }
 
-    public String getFragmentPlan(String sql) throws Exception {
-        String s = UtFrameUtils.getPlanAndFragment(connectContext, sql).second.
-                getExplainString(TExplainLevel.NORMAL);
-        return s;
-    }
-
-    private Table getTable(String dbName, String mvName) {
-        Database db = GlobalStateMgr.getCurrentState().getDb(dbName);
-        Table table = db.getTable(mvName);
-        Assert.assertNotNull(table);
-        return table;
-    }
-
-    private MaterializedView getMv(String dbName, String mvName) {
-        Table table = getTable(dbName, mvName);
-        Assert.assertTrue(table instanceof MaterializedView);
-        MaterializedView mv = (MaterializedView) table;
-        return mv;
-    }
-
-    private void refreshMaterializedView(String dbName, String mvName) throws SQLException {
-        cluster.runSql(dbName, String.format("refresh materialized view %s with sync mode", mvName));
-    }
-
-    private void createAndRefreshMv(String dbName, String mvName, String sql) throws Exception {
-        starRocksAssert.withMaterializedView(sql);
-        cluster.runSql(dbName, String.format("refresh materialized view %s with sync mode", mvName));
-    }
-
-    private void dropMv(String dbName, String mvName) throws Exception {
-        starRocksAssert.dropMaterializedView(mvName);
-    }
-
-    public static OptExpression getOptimizedPlan(String sql, ConnectContext connectContext) {
-        StatementBase mvStmt;
-        try {
-            List<StatementBase> statementBases =
-                    com.starrocks.sql.parser.SqlParser.parse(sql, connectContext.getSessionVariable());
-            Preconditions.checkState(statementBases.size() == 1);
-            mvStmt = statementBases.get(0);
-        } catch (ParsingException parsingException) {
-            return null;
-        }
-        Preconditions.checkState(mvStmt instanceof QueryStatement);
-        Analyzer.analyze(mvStmt, connectContext);
-        QueryRelation query = ((QueryStatement) mvStmt).getQueryRelation();
-        ColumnRefFactory columnRefFactory = new ColumnRefFactory();
-        LogicalPlan logicalPlan =
-                new RelationTransformer(columnRefFactory, connectContext).transformWithSelectLimit(query);
-        Optimizer optimizer = new Optimizer();
-        return optimizer.optimize(
-                connectContext,
-                logicalPlan.getRoot(),
-                new PhysicalPropertySet(),
-                new ColumnRefSet(logicalPlan.getOutputColumn()),
-                columnRefFactory);
-    }
-
-    public List<PhysicalScanOperator> getScanOperators(OptExpression root, String name) {
-        List<PhysicalScanOperator> results = Lists.newArrayList();
-        getScanOperators(root, name, results);
-        return results;
-    }
-
-    private void getScanOperators(OptExpression root, String name, List<PhysicalScanOperator> results) {
-        if (root.getOp() instanceof PhysicalScanOperator
-                && ((PhysicalScanOperator) root.getOp()).getTable().getName().equals(name)) {
-            results.add((PhysicalScanOperator) root.getOp());
-        }
-        for (OptExpression child : root.getInputs()) {
-            getScanOperators(child, name, results);
-        }
-    }
-
-    @Test
-    public void testNestedMVs1() throws Exception {
-        createAndRefreshMv("test", "nested_mv1", "create materialized view nested_mv1 " +
-                " distributed by hash(empid) as" +
-                " select * from emps;");
-        createAndRefreshMv("test", "nested_mv2", "create materialized view nested_mv2 " +
-                " distributed by hash(empid) as" +
-                " select empid, sum(deptno) from emps group by empid;");
-        createAndRefreshMv("test", "nested_mv3", "create materialized view nested_mv3 " +
-                " distributed by hash(empid) as" +
-                " select * from nested_mv2 where empid > 1;");
-        String plan = getFragmentPlan("select empid, sum(deptno) from emps where empid > 1 group by empid");
-        System.out.println(plan);
-        Assert.assertTrue(plan.contains("0:OlapScanNode\n" +
-                "     TABLE: nested_mv3\n" +
-                "     PREAGGREGATION: ON\n" +
-                "     partitions=1/1\n" +
-                "     rollup: nested_mv3"));
-        dropMv("test", "nested_mv1");
-        dropMv("test", "nested_mv2");
-        dropMv("test", "nested_mv3");
-    }
-
-    @Test
-    public void testExternalNestedMVs1() throws Exception {
-        connectContext.getSessionVariable().setEnableMaterializedViewUnionRewrite(true);
-        createAndRefreshMv("test", "hive_nested_mv_1",
-                "create materialized view hive_nested_mv_1 distributed by hash(s_suppkey) " +
-                        "PROPERTIES (\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\"\n" +
-                        ") " +
-                        " as select s_suppkey, s_name, s_address, s_acctbal from hive0.tpch.supplier");
-        createAndRefreshMv("test", "hive_nested_mv_2",
-                "create materialized view hive_nested_mv_2 distributed by hash(s_suppkey) " +
-                        "PROPERTIES (\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\"\n" +
-                        ") " +
-                        " as select s_suppkey, sum(s_acctbal) from hive0.tpch.supplier group by s_suppkey");
-        createAndRefreshMv("test", "hive_nested_mv_3",
-                "create materialized view hive_nested_mv_3 distributed by hash(s_suppkey) " +
-                        "PROPERTIES (\n" +
-                        "\"force_external_table_query_rewrite\" = \"true\"\n" +
-                        ") " +
-                        " as select * from hive_nested_mv_2 where s_suppkey > 1");
-        String query1 = "select s_suppkey, sum(s_acctbal) from hive0.tpch.supplier where s_suppkey > 1 group by s_suppkey ";
-        String plan1 = getFragmentPlan(query1);
-        System.out.println(plan1);
-        Assert.assertTrue(plan1.contains("0:OlapScanNode\n" +
-                "     TABLE: hive_nested_mv_3\n" +
-                "     PREAGGREGATION: ON\n"));
-        dropMv("test", "hive_nested_mv_1");
-        dropMv("test", "hive_nested_mv_2");
-        dropMv("test", "hive_nested_mv_3");
-    }
-
     @Test
     public void testTabletHintForbidMvRewrite() throws Exception {
         createAndRefreshMv("test", "forbid_mv_1", "create materialized view forbid_mv_1" +
@@ -2649,67 +1433,5 @@ public class MvRewriteOptimizationTest {
                 "     TABLE: test_all_type\n" +
                 "     PREAGGREGATION: ON");
         dropMv("test", "forbid_mv_1");
-    }
-
-    @Test
-    public void testNullPartitionRewriteWithLoad() throws Exception {
-        {
-            cluster.runSql("test", "insert into test_base_part values(1, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(100, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(200, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(1000, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(2000, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(2500, 1, 2, 3)");
-
-            createAndRefreshMv("test", "partial_mv_12", "CREATE MATERIALIZED VIEW `partial_mv_12`\n" +
-                    "COMMENT \"MATERIALIZED_VIEW\"\n" +
-                    "PARTITION BY (`c3`)\n" +
-                    "DISTRIBUTED BY HASH(`c1`) BUCKETS 6\n" +
-                    "REFRESH MANUAL\n" +
-                    "PROPERTIES (\n" +
-                    "\"replication_num\" = \"1\",\n" +
-                    "\"storage_medium\" = \"HDD\"\n" +
-                    ")\n" +
-                    "AS SELECT `c1`, `c3`, sum(`c4`) AS `total`\n" +
-                    "FROM `test_base_part`\n" +
-                    "WHERE `c3` is null\n" +
-                    "GROUP BY `c3`, `c1`;");
-
-            String query = "select c1, c3, sum(c4) from test_base_part group by c1, c3;";
-            String plan = getFragmentPlan(query);
-            PlanTestBase.assertContains(plan, "partial_mv_12", "PREDICATES: 10: c3 IS NOT NULL");
-            starRocksAssert.dropMaterializedView("partial_mv_12");
-        }
-
-        {
-            cluster.runSql("test", "insert into test_base_part values(1, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(100, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(200, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(1000, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(2000, 1, 2, 3)");
-            cluster.runSql("test", "insert into test_base_part values(2500, 1, 2, 3)");
-
-            createAndRefreshMv("test", "partial_mv_13", "CREATE MATERIALIZED VIEW `partial_mv_13`\n" +
-                    "COMMENT \"MATERIALIZED_VIEW\"\n" +
-                    "PARTITION BY (`c3`)\n" +
-                    "DISTRIBUTED BY HASH(`c1`) BUCKETS 6\n" +
-                    "REFRESH MANUAL\n" +
-                    "PROPERTIES (\n" +
-                    "\"replication_num\" = \"1\",\n" +
-                    "\"storage_medium\" = \"HDD\"\n" +
-                    ")\n" +
-                    "AS SELECT `c1`, `c3`, sum(`c4`) AS `total`\n" +
-                    "FROM `test_base_part`\n" +
-                    "WHERE `c3` is null\n" +
-                    "GROUP BY `c3`, `c1`;");
-
-            // test update for null partition
-            cluster.runSql("test", "insert into test_base_part partition(p1) values(null, 1, null, 3)");
-
-            String query = "select c1, c3, sum(c4) from test_base_part group by c1, c3;";
-            String plan = getFragmentPlan(query);
-            PlanTestBase.assertNotContains(plan, "partial_mv_13");
-            starRocksAssert.dropMaterializedView("partial_mv_13");
-        }
     }
 }
